perf: Allow `secret` and `publicKey` options to be `crypto.KeyObject`
When JwtService is initialized with `publicKey` as a string or Buffer,
`verify()` and `verifyAsync()` pass it to "jsonwebtoken.verify()",
which creates an instance of `crypto.KeyObject` from it via
`crypto.createPublicKey()`. This is not free. Initializing `publicKey`
with a `KeyObject` avoids this transformation in "jsonwebtoken". On my
laptop, it makes `verify` twice faster.

The same goes for `secret`, used in `sign()`, `verify()` and their
asynchronous variants. Initializing with a `KeyObject` (built via
`crypto.createSecretKey`) makes these functions ~50 times faster.

See also auth0/node-jsonwebtoken@966.